add IE11 support with a Javascript fallback built against KTX-Software 78790b5
